Support im.list + im.history. lots can be reused from channels. Create the 'Common' module.
